requirements.txt
jieba
protobuf
grpcio-tools
python-snappy
